Add ability to insert media directly from Ace editor
Added option to new Ace editor custom toolbar that lets the user insert
HTML into the Ace editor via selecting a media item from the media
library. The editor will try it's best to add the appropriate HTML for
the current selection in the Ace session. If the user currently has the
cursor within the `src`/`href` attribute then it'll only add the URL,
otherwise it will try to add the whole element.

Currently only documents & images are supported.