Improve JS structure to improve scaling out
As we add more options to the editor, currently it's being added to a
single JS file (`orchard.editor.js`). To isolate individual features
to single files we've introduced a pattern for adding "plugins". These
should be added to the `plugins` directory. Each plugin will have an
`action` and a function (`exec`) that'll be executed when a toolbar
item that has a matching action is interacted with.

Currently only the `insert-media` action has been moved across.