Improve visual editor loading experience
When the editor loads the user will see the code editor then the
visual editor will load in. To improve this the code editor no longer
displays while the visual editor is being loaded in. Also the toolbar
items don't show until the visual editor has loaded.

These changes have resulted in a faster loading experience as there
is no longer a 2 second wait to load the visual editor.