Improve media picker logic within visual editor
When attemtping to add media within the visual editor, currently it'll
loaded the Orchard media picker within the visual editor `iframe`. The
downside to this is it would be restricted by the dimensions of the `iframe`,
which can sometimes be very small. To improve this, it needed to have a
similiar experience to the code editor, where the media picker would
take up all the available space within the browser window (except when
used within Layouts).

This update will refactor how the media picker is loaded. The visual
editor will instead send a message to the main editor, which would handle
loading the media picker. Once the media is selected, a message is
sent to the visual editor.