Fix rare issue with editor styles not loading
Fix an issue that could cause the editor styles to not load that would
occur when orchard mistakenly used a Styles.css from another module
instead of the one defined by this module.

Renamed the stylesheet to Moov2.Editor.css to prevent this from occuring
if another module uses Styles.css.